+++ tests/conftest.py
@@ -2,6 +2,7 @@ from datetime import date
 
 import pytest
 from django import VERSION as DJANGO_VERSION
+from django.db.models import CharField
 from mock import Mock
 
 from django.db.models import Value
@@ -16,6 +17,16 @@ from .dummy_lib.models import ReleaseTypeModel
 collect_ignore = []
 if DJANGO_VERSION < (4, 1):
     collect_ignore.append('test_queries/test_async.py')
+
+
+@pytest.fixture(autouse=True, scope='session')
+def sha1_lookup():
+    try:
+        from django.db.models.functions import SHA1
+    except ImportError:
+        pass
+    else:
+        CharField.register_lookup(SHA1)
 
 
 @pytest.fixture